pkcs11: do not assume c_ulong is u64
Instead use macros to check size of pointers, and thereby determine if
we are on 32bit or 64bit platform.  On 32bit platforms c_ulong is u32,
and otherwise u64.

This works well for android, but I suppose there might be other exotic
platforms where size of c_ulong and size of pointers do not
necessarily follow each other.